Updated the SystemFileStream class to always own the SystemFile
The SystemFileStream has also been simplified. The fake length and base
offset to seek within the SystemFile has been removed and a constructor
which accepts a path and an open mode which mirrors the Open() function.
