+++ Code/Framework/AzCore/AzCore/IO/FileIO.cpp
@@ -143,149 +143,6 @@ namespace AZ::IO
         return SEEK_SET;
     }
 
-    void UpdateOpenModeForReading(OpenMode& openMode)
-    {
-        if (AnyFlag(openMode & OpenMode::ModeRead))
-        {
-            if (AnyFlag(openMode & OpenMode::ModeText))
-            {
-                OpenMode extraModes = openMode & (OpenMode::ModeUpdate | OpenMode::ModeAppend);
-                openMode = OpenMode::ModeRead | OpenMode::ModeBinary | extraModes;
-            }
-            else if (!AnyFlag(openMode & OpenMode::ModeBinary))
-            {
-                // if you haven't supplied any flag, supply binary
-                openMode = openMode | OpenMode::ModeBinary;
-            }
-        }
-    }
-
-    OpenMode GetOpenModeFromStringMode(const char* mode)
-    {
-        OpenMode openMode = OpenMode::Invalid;
-
-        if (strstr(mode, "w"))
-        {
-            openMode |= OpenMode::ModeWrite;
-        }
-
-        if (strstr(mode, "r"))
-        {
-            openMode |= OpenMode::ModeRead;
-        }
-
-        if (strstr(mode, "a"))
-        {
-            openMode |= OpenMode::ModeAppend;
-        }
-
-        if (strstr(mode, "b"))
-        {
-            openMode |= OpenMode::ModeBinary;
-        }
-
-        if (strstr(mode, "t"))
-        {
-            openMode |= OpenMode::ModeText;
-        }
-
-        if (strstr(mode, "+"))
-        {
-            openMode |= OpenMode::ModeUpdate;
-        }
-
-        UpdateOpenModeForReading(openMode);
-
-        return openMode;
-    }
-
-    const char* GetStringModeFromOpenMode(OpenMode mode)
-    {
-        UpdateOpenModeForReading(mode);
-        // Append is highest priority, followed by write and then read
-        // APPEND
-        if (AnyFlag(mode & OpenMode::ModeAppend))
-        {
-            if (AnyFlag(mode & OpenMode::ModeUpdate))
-            {
-                if (AnyFlag(mode & OpenMode::ModeBinary))
-                {
-                    return "a+b";
-                }
-                if (AnyFlag(mode & OpenMode::ModeText))
-                {
-                    return "a+t";
-                }
-                return "a+";
-            }
-            if (AnyFlag(mode & OpenMode::ModeBinary))
-            {
-                return "ab";
-            }
-            if (AnyFlag(mode & OpenMode::ModeText))
-            {
-                return "at";
-            }
-            return "a";
-        }
-
-        // WRITE
-        if (AnyFlag(mode & OpenMode::ModeWrite))
-        {
-            if (AnyFlag(mode & OpenMode::ModeUpdate))
-            {
-                if (AnyFlag(mode & OpenMode::ModeBinary))
-                {
-                    return "w+b";
-                }
-                if (AnyFlag(mode & OpenMode::ModeText))
-                {
-                    return "w+t";
-                }
-                return "w+";
-            }
-            if (AnyFlag(mode & OpenMode::ModeBinary))
-            {
-                return "wb";
-            }
-            if (AnyFlag(mode & OpenMode::ModeText))
-            {
-                return "wt";
-            }
-            return "w";
-        }
-
-        // READ
-        if (AnyFlag(mode & OpenMode::ModeRead))
-        {
-            if (AnyFlag(mode & OpenMode::ModeUpdate))
-            {
-                if (AnyFlag(mode & OpenMode::ModeBinary))
-                {
-                    return "r+b";
-                }
-                if (AnyFlag(mode & OpenMode::ModeText))
-                {
-                    return "r+t";
-                }
-                return "r+";
-            }
-            if (AnyFlag(mode & OpenMode::ModeBinary))
-            {
-                return "rb";
-            }
-            if (AnyFlag(mode & OpenMode::ModeText))
-            {
-                return "rt";
-            }
-            return "r";
-        }
-
-        // Bad open mode passed in
-        AZ_Error("FileIO", false, "A bad open mode was sent to GetStringModeFromOpenMode()");
-        return "";
-    }
-
     bool NameMatchesFilter(AZStd::string_view name, AZStd::string_view filter)
     {
         return AZStd::wildcard_match(filter, name);
