+++ Code/Framework/AzCore/AzCore/IO/FileIO.h
@@ -46,36 +46,6 @@ namespace AZ
         SeekType GetSeekTypeFromFSeekMode(int mode);
         int GetFSeekModeFromSeekType(SeekType type);
 
-        enum class OpenMode : AZ::u32
-        {
-            Invalid = 0,
-            ModeRead = (1 << 0),
-            ModeWrite = (1 << 1),
-            ModeAppend = (1 << 2),
-            ModeBinary = (1 << 3),
-            ModeText = (1 << 4),
-            ModeUpdate = (1 << 5),
-            ModeCreatePath = (1 << 6),
-        };
-
-        inline bool AnyFlag(OpenMode a)
-        {
-            return a != OpenMode::Invalid;
-        }
-
-        AZ_DEFINE_ENUM_BITWISE_OPERATORS(OpenMode)
-
-        OpenMode GetOpenModeFromStringMode(const char* mode);
-
-        const char* GetStringModeFromOpenMode(OpenMode mode);
-
-        // when reading, we ignore text mode requests, becuase text mode will not be supported from reading from
-        // odd devices such as pack files, remote reads, and within compressed volumes anyway.  This makes it behave the same as
-        // pak mode as it behaves loose.  In practice, developers should avoid depending on text mode for reading, specifically depending
-        // on the platform-specific concept of removing '\r' characters from a text stream.
-        // having the OS swallow characters without letting us know also messes up things like lookahead cache, optizing functions like FGetS and so on.
-        void UpdateOpenModeForReading(OpenMode& mode);
-
         enum class ResultCode : AZ::u32
         {
             Success = 0,
