+++ Code/Framework/AzToolsFramework/AzToolsFramework/UI/LegacyFramework/Core/EditorFrameworkApplication.cpp
@@ -362,11 +362,10 @@ namespace LegacyFramework
         qstrcpy(m_applicationFilePath, applicationFilePath.c_str());
 
         // load all application entities, if present:
-        AZ::IO::SystemFile cfg;
 
-        if (cfg.Open(m_applicationFilePath, AZ::IO::SystemFile::SF_OPEN_READ_ONLY))
+        if (AZ::IO::SystemFileStream stream(m_applicationFilePath, AZ::IO::OpenMode::ModeRead);
+            stream.IsOpen())
         {
-            AZ::IO::SystemFileStream stream(&cfg, false);
             stream.Seek(0, AZ::IO::GenericStream::ST_SEEK_BEGIN);
             AZ::ObjectStream::LoadBlocking(&stream, *GetSerializeContext(),
                 [this](void* classPtr, const AZ::Uuid& classId, const AZ::SerializeContext* sc)
@@ -377,7 +376,6 @@ namespace LegacyFramework
                         m_applicationEntity = entity;
                     }
                 });
-            cfg.Close();
         }
 
         if (!m_applicationEntity)
@@ -456,7 +454,7 @@ namespace LegacyFramework
             AZ_Warning("ComponentApplication", entityWriteOk, "Failed to write application entity to application file %s!", applicationFilePath.c_str());
             bool flushOk = objStream->Finalize();
             AZ_Warning("ComponentApplication", flushOk, "Failed finalizing application file %s!", applicationFilePath.c_str());
-            
+
             if (entityWriteOk && flushOk)
             {
                 if (IO::SystemFile::Rename(tmpFileName.c_str(), applicationFilePath.c_str(), true))
